Eine Zahl zu breit -- und das 37. Tor, das es misst
CI-FIX ZU RUNDE 196. ModelArmorRPA liess den Luefter der Panzerruestung mit
Axis.ZP.rotationDegrees(-(System.currentTimeMillis() / 2D % 360D)) drehen.
rotationDegrees nimmt ein float; der Ausdruck ist ein double. javac lehnt das
ab ("possible lossy conversion from double to float"), und keines der
sechsunddreissig Tore sah es: die Signatur der gerufenen Methode steht in der
Bibliothek, nicht im Port, und syntax-check uebersetzt ohne Minecraft-
Klassenpfad. Ein Cast behebt es.

DAS IST DER ZWEITE CI-LAUF IN VIER RUNDEN, den eine reine Typfrage kostet --
Runde 193 war die Stelligkeit eines Konstruktors, diese hier die Breite einer
Zahl. Beide Male stand die Antwort im Projekt selbst, nur sah niemand hin.

DAS 37. TOR: tools/narrow-check.sh, nach dem Mehrheitsprinzip von
signature-check.sh. Uebergibt eine Aufrufstelle ein double, wo alle anderen
Aufrufstellen derselben Methode desselben Empfaengertyps ein float uebergeben,
ist sie es, die falsch ist. Bei Axis.rotationDegrees stehen 1090 Belege gegen
diesen einen.

VIER BEDINGUNGEN, und jede war noetig. Der Empfaengertyp gehoert in den
Schluessel: ueber den nackten Namen gemessen kollidieren Vec3.scale(double) und
RenderContext.scale(float), was zwoelf Fehlalarme gab. Mindestens acht Belege,
sonst ist "die Mehrheit" nichts wert. Hoechstens ein Zehntel Abweichler, sonst
ist es eine Ueberladung, die beides nimmt. Und nur zweifelsfreie Ausdruecke --
was ein Cast beruehrt, zaehlt gar nicht.

GRUPPIERUNGSKLAMMERN SIND KEINE AUFRUFKLAMMERN, und der erste Entwurf
verwechselte beides: -(a / 2D) ist ein double, obwohl die Zahl in Klammern
steht, foo(2D) dagegen nicht. So verdeckt gemessen fand das Tor den Fehler
nicht, gegen den es gebaut wurde. Die zweite Fassung unterscheidet danach, ob
vor der Klammer ein Name steht.

Gemessen: 38980 Aufrufstellen mit Empfaengertyp, null Funde. Nimmt man den Cast
in ModelArmorRPA wieder weg, meldet das Tor genau diese Zeile -- und sonst
nichts.

Alle 37 Tore gruen.

+++ src/main/java/com/hbm/render/model/armor/ModelArmorRPA.java
@@ -106,7 +106,7 @@ public class ModelArmorRPA extends ModelArmorBase {
         if(this.body.xRot != 0F) RenderContext.mulPose(Axis.XP.rotation(this.body.xRot));
 
         RenderContext.translate(0F, 4.875F * px, 0F);
-        RenderContext.mulPose(Axis.ZP.rotationDegrees(-(System.currentTimeMillis() / 2D % 360D)));
+        RenderContext.mulPose(Axis.ZP.rotationDegrees((float) -(System.currentTimeMillis() / 2D % 360D)));
         RenderContext.translate(0F, -4.875F * px, 0F);
 
         /* Der Luefter ist schon an seinem Platz -- sein eigener Drehpunkt bleibt bei null. */
